Fixed Timer so value should be retrievable
Made TimerApp class within quizpage.js that acts as the holder of the Time Value. The Timer.js file then receives the time from said class and on interval calls the TimerApp function that reduces the timeValue by 1. This timeValue should now be able to be manipulated within quizPage.js